% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handles.r
\name{handles}
\alias{handles}
\alias{get.default.handle}
\alias{set.default.handle}
\title{Default Handles}
\usage{
get.default.handle()

set.default.handle(handle)
}
\arguments{
\item{handle}{The return from \code{lldb.load()}.}
}
\value{
\code{get.default.handle()} returns the default handle if set, and errors
otherwise. \code{set.default.handle()} sets the global default handle value.
}
\description{
Functions to get and set the default handle.
}
\seealso{
\code{\link{lldb.load}}
}
