% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.r
\name{lldb.attach}
\alias{lldb.attach}
\title{lldb.attach}
\usage{
lldb.attach(pid = NULL, quiet = FALSE, setdefault = TRUE)
}
\arguments{
\item{pid}{the PID of the process}

\item{setdefault}{if \code{TRUE}, then the default handle will be updated with the return.}
}
\value{
An object of class \code{lldb_handle}; an external pointer.
}
\description{
Attaches the debugger to an existing process. This would replace a call to lldb.load().
}
\seealso{
\code{\link{lldb.breakfun}} \code{\link{lldb.breakline}} \code{\link{lldb.load}}
}
