% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.r
\name{lldb.break}
\alias{lldb.break}
\title{lldb.break}
\usage{
lldb.break(handle, file, line)
}
\arguments{
\item{handle}{handle returned from lldb.load}

\item{file}{source file to break in}

\item{line}{line number to break at}
}
\value{
An invisible return code.
}
\description{
Set a breakpoint.
}
\examples{
\dontrun{
handle <- lldb.load("/path/to/binary")

### Break at line 10 of the specified source file 
lldb.break(handle, "/path/to/source.c", 10)
}

}
\seealso{
\code{\link{lldb.run}}
}

