% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.r
\name{lldb.breakfun}
\alias{lldb.breakfun}
\title{lldb.breakfun}
\usage{
lldb.breakfun(symbol = NULL, module = NULL, handle = NULL)
}
\arguments{
\item{symbol}{symbol (function) to break in}

\item{module}{module to search for symbol}

\item{handle}{handle returned from \code{lldb.load()} or \code{NULL} for the default handle}
}
\value{
An invisible return code.
}
\description{
Set a breakpoint at the specified symbol
}
\examples{
\dontrun{
library(lldbR)
lldb.load("/path/to/binary")

### Break in main()
lldb.breakfun(symbol="main")
}

}
\seealso{
\code{\link{lldb.run}} \code{\link{lldb.breakline}}
}

