% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.r
\name{lldb.breakline}
\alias{lldb.breakline}
\title{lldb.breakline}
\usage{
lldb.breakline(file = NULL, line = 0, handle = NULL)
}
\arguments{
\item{file}{source file to break in}

\item{line}{line number to break at}

\item{handle}{handle returned from \code{lldb.load()} or \code{NULL} for the default handle}
}
\value{
An invisible return code.
}
\description{
Set a breakpoint at the specified line number
}
\examples{
\dontrun{
library(lldbR)
lldb.load("/path/to/binary")

### Break at line 10 of the specified source file 
lldb.break(file="/path/to/source.c", line=10)
}

}
\seealso{
\code{\link{lldb.run}} \code{\link{lldb.breakfun}}
}

