% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.r
\name{lldb.continue}
\alias{lldb.continue}
\title{lldb.continue}
\usage{
lldb.continue(handle = NULL)
}
\arguments{
\item{handle}{handle returned from \code{lldb.load()} or \code{NULL} for the default handle}
}
\value{
An invisible return code.
}
\description{
Continue process execution until the next breakpoint or exit.
}
\seealso{
\code{\link{lldb.breakline}}\code{\link{lldb.breakfun}}
}
