% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.r
\name{lldb.exit}
\alias{lldb.exit}
\title{lldb.exit}
\usage{
lldb.exit(handle = NULL)
}
\arguments{
\item{handle}{handle returned from \code{lldb.load()} or \code{NULL} for the default handle}
}
\description{
Halt the process, destroy the handle, and reset the default
handle as necessary.
}
\details{
Strictly speaking, you do not need to call this function
as it will be called automatically during garbage collection.
However, if you wish to halt the attached process manually,
this is the simplest way to do so.
}
\examples{
\dontrun{
library(lldbR)
lldb.load("/path/to/binary")
lldb.exit()
}

}
\seealso{
\code{\link{handles}, \link{lldb.load}}
}
