% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.r
\name{lldb.expr}
\alias{lldb.expr}
\title{lldb.expr}
\usage{
lldb.expr(handle, expr, offset, size)
}
\arguments{
\item{handle}{handle returned from lldb.load}

\item{expr}{expression to evaluate}

\item{offset}{offset index to read from}

\item{size}{number of elements to read}
}
\value{
The requested data.
}
\description{
Extract data from the process into R via an expression.
}
\examples{
\dontrun{
library(lldbR)
handle <- lldb.load("/path/to/binary")

### Break at line 10 of the specified source file 
lldb.break(handle, "/path/to/source.c", 10)
lldb.run(handle)

### Extract x[0]
lldb.expr(handle, "x")
### Extract x[2] to x[4]
lldb.expr(handle, "x", 2, 3)
}

}

