% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.r
\name{lldb.run}
\alias{lldb.run}
\title{lldb.run}
\usage{
lldb.run(handle, args)
}
\arguments{
\item{handle}{handle returned from lldb.load}

\item{args}{command line arguments for the process}
}
\description{
lldb.run
}

