% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.r
\name{lldb.step}
\alias{lldb.step}
\alias{lldb.stepto}
\title{Step functions}
\usage{
lldb.step(method = "into", handle = NULL)

lldb.stepto(file = NULL, line = 0, handle = NULL)
}
\arguments{
\item{method}{one of c("out","over","into")}

\item{handle}{handle returned from \code{lldb.load()} or \code{NULL} for the default handle}

\item{file}{source file to step to}

\item{line}{line number to step to}
}
\value{
An invisible return code.
}
\description{
Continue process execution until the next line

lldb.step steps either 'into', 'out' of, or 'over' a function or line

lldb.stepto steps until reaching the specified line
}

