% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lldbR.r
\docType{package}
\name{lldbR}
\alias{lldbR}
\alias{lldbR-package}
\title{LLDB API R wrapper}
\description{
lldbR
}
\details{
LLDB API R wrapper
}
\author{
Christian Heckendorf \email{heckendorfc@gmail.com}
}
\keyword{package}
