% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{splitter}
\alias{splitter}
\title{Character Splitter}
\usage{
splitter(x)
}
\arguments{
\item{x}{string containing to be processed and split}
}
\value{
The string split into a character vector
}
\description{
A utility function that splits a string into a character vector suitable for
the args parameter of lldb.run
}
\seealso{
\code{\link{lldb.run}}
}

