% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.r
\name{lldb.break}
\alias{lldb.break}
\title{lldb.break}
\usage{
lldb.break(handle, file = NULL, line = 0, symbol = NULL, module = NULL)
}
\arguments{
\item{handle}{handle returned from lldb.load}

\item{file}{source file to break in}

\item{line}{line number to break at}

\item{symbol}{symbol (function) to break in}

\item{module}{module to search for symbol}
}
\value{
An invisible return code.
}
\description{
Set a breakpoint. Requires either file+line or symbol+(optional)module.
}
\examples{
\dontrun{
handle <- lldb.load("/path/to/binary")

### Break at line 10 of the specified source file 
lldb.break(handle, file="/path/to/source.c", line=10)
### Break in main()
lldb.break(handle, symbol="main")
}

}
\seealso{
\code{\link{lldb.run}}
}

