% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.r
\name{lldb.break}
\alias{lldb.break}
\title{lldb.break}
\usage{
lldb.break(file, line, handle = NULL)
}
\arguments{
\item{file}{source file to break in}

\item{line}{line number to break at}

\item{handle}{handle returned from \code{lldb.load()} or \code{NULL} for the default handle}
}
\value{
An invisible return code.
}
\description{
Set a breakpoint.
}
\examples{
\dontrun{
library(lldbR)
lldb.load("/path/to/binary")

### Break at line 10 of the specified source file 
lldb.break("/path/to/source.c", 10)
}

}
\seealso{
\code{\link{lldb.run}}
}

