% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break.r
\name{lldb.break}
\alias{lldb.break}
\alias{lldb.breakfun}
\alias{lldb.breakline}
\title{Set breakpoints}
\usage{
lldb.breakline(file = NULL, line = 0, handle = NULL)

lldb.breakfun(symbol = NULL, module = NULL, handle = NULL)
}
\arguments{
\item{file}{source file to break in}

\item{line}{line number to break at}

\item{handle}{handle returned from \code{lldb.load()} or \code{NULL} for the default handle}

\item{symbol}{symbol (function) to break in}

\item{module}{module to search for symbol}
}
\value{
An invisible return code.
}
\description{
Stop execution at certain posititions in the source

lldb.breakline sets a breakpoint at the specified line

lldb.breakfun sets a breakpoint at the specified symbol
}
\examples{
\dontrun{
library(lldbR)
lldb.load("/path/to/binary")

### Break at line 10 of the specified source file 
lldb.breakline(file="/path/to/source.c", line=10)

### Break in main()
lldb.breakfun(symbol="main")
}

}
\seealso{
\code{\link{lldb.run}}
}

