% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.r
\name{lldb.continue}
\alias{lldb.continue}
\title{lldb.continue}
\usage{
lldb.continue(handle)
}
\arguments{
\item{handle}{handle returned from lldb.load}
}
\value{
An invisible return code.
}
\description{
Continue process execution until the next breakpoint or exit.
}
\seealso{
\code{\link{lldb.break}}
}

