% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.r
\name{lldb.expr}
\alias{lldb.expr}
\title{lldb.expr}
\usage{
lldb.expr(handle, expr, offset, size)
}
\arguments{
\item{handle}{handle returned from lldb.load}

\item{expr}{expression to evaluate}

\item{offset}{offset index to read from}

\item{size}{number of elements to read}
}
\description{
lldb.expr
}

