% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.r
\name{lldb.expr}
\alias{lldb.expr}
\title{lldb.expr}
\usage{
lldb.expr(expr, offset = 0, size = 1, fixed = TRUE, handle = NULL)
}
\arguments{
\item{expr}{expression to evaluate}

\item{offset}{offset index to read from}

\item{size}{number of elements to read}

\item{fixed}{if true, the expr parameter is a fixed string and should not be evaluated
for special sequences}

\item{handle}{handle returned from \code{lldb.load()} or \code{NULL} for the default handle}
}
\value{
The requested data.
}
\description{
Extract data from the process into R via an expression.
}
\details{
Expressions containing the special sequence "$$" will iterate over array at
that position. This allows for extraction of child elements in an array of
structs.
}
\examples{
\dontrun{
library(lldbR)
lldb.load("/path/to/binary")

### Break at line 10 of the specified source file 
lldb.breakline("/path/to/source.c", 10)
lldb.run()

### Extract x[0]
lldb.expr("x")
### Extract x[2] to x[4]
lldb.expr("x", 2, 3)
### Extract y[2].start to y[4].start
lldb.expr("y[$$].start", 2, 3, fixed=FALSE)
}

}
