% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.r
\name{lldb.expr}
\alias{lldb.expr}
\title{lldb.expr}
\usage{
lldb.expr(expr, offset = 0, size = 1, handle = NULL)
}
\arguments{
\item{expr}{expression to evaluate}

\item{offset}{offset index to read from}

\item{size}{number of elements to read}

\item{handle}{handle returned from \code{lldb.load()} or \code{NULL} for the default handle}
}
\value{
The requested data.
}
\description{
Extract data from the process into R via an expression.
}
\examples{
\dontrun{
library(lldbR)
lldb.load("/path/to/binary")

### Break at line 10 of the specified source file 
lldb.break("/path/to/source.c", 10)
lldb.run()

### Extract x[0]
lldb.expr("x")
### Extract x[2] to x[4]
lldb.expr("x", 2, 3)
}

}

