% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.r
\name{lldb.load}
\alias{lldb.load}
\title{lldb.load}
\usage{
lldb.load(args)
}
\arguments{
\item{process}{process name}
}
\value{
An object of class \code{lldb_handle}; an external pointer.
}
\description{
The initializer that creates a "handle" for lldb.
}
\examples{
\dontrun{
handle <- lldb.load("/path/to/binary")
handle
}

}
\seealso{
\code{\link{lldb.expr}}
}

