% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.r
\name{lldb.run}
\alias{lldb.run}
\title{lldb.run}
\usage{
lldb.run(args = NULL, handle = NULL)
}
\arguments{
\item{args}{command line arguments for the process or \code{NULL} (the default) for no
arguments. Each argument should be in its own string. \code{splitter} can be
used to split a single string.}

\item{handle}{handle returned from \code{lldb.load()} or \code{NULL} for the default handle}
}
\value{
An invisible return code.
}
\description{
Runs the process.
}
\seealso{
\code{\link{lldb.break}} \code{\link{splitter}}
}

